+++ leeger/util/excel.py
@@ -87,6 +87,9 @@ def leagueToExcel(league: League, filePath: str, **kwargs) -> None:
                   ref="A1:" + get_column_letter(worksheet.max_column) + str(worksheet.max_row))
     worksheet.add_table(table)
 
+    # freeze owner name column
+    worksheet.freeze_panes = "B1"
+
     # save
     workbook.save(filePath)
 
@@ -169,6 +172,8 @@ def yearToExcel(year: Year, filePath: str, **kwargs) -> None:
     table = Table(displayName=f"YearStats{year.yearNumber}",
                   ref="A1:" + get_column_letter(worksheet.max_column) + str(worksheet.max_row))
     worksheet.add_table(table)
+    # freeze team name column
+    worksheet.freeze_panes = "B1"
 
     # save
     workbook.save(filePath)